:recycle: (refactor) proper folder structure for the interpreter
