+++ libs/functions/chatbot/src/lib/services/interpreter/received-message-interpreter/receive-interpreter.factory.ts
@@ -1,5 +1,5 @@
 import { Platforms } from "@app/model/convs-mgr/conversations/admin/system";
-import { WhatsappReceiveMessageInterpreter } from "./whatsapp-api-message-to-base-message.class";
+import { WhatsappReceiveMessageInterpreter } from "../whatsapp-api-message-to-base-message.class";
 
 /** Resolves the receive message interpreter based on the Platform */
 export class ReceiveInterpreterFactory {
