+++ libs/functions/chatbot/src/lib/services/interpreter/received-message-interpreter/receive-message-interpreter-abstract.class.ts
