+++ libs/functions/chatbot/src/lib/services/interpreter/send-message-interpreter/send-message-interpreter-abstract.class.ts
